backend/pyproject.toml
[project]
name = "bbs-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.135.3",
    "pydantic-settings>=2.7.1",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.5",
    "pyjwt>=2.12.1",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.34.0",
]
